WoT restructure A4: cellAccess + cellCoords -> cellLocator/ (git-mv + imports)
